--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@ docker-compose.override.yaml
 .tmp
 
 perf.data*
+flamegraph.svg