@MrCroxx 2024-04-18
## Benchmark

### cmd

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 204800 --region-size 64 --lookup-range 10000 --flushers 12 --reclaimers 12 --time 180 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --io-size 262144 --distribution zipf --distribution-zipf-s 0.8
```

### main

```
Total:
disk total iops: 3822.8
disk total throughput: 483.8 MiB/s
disk read iops: 356.3
disk read throughput: 23.7 MiB/s
disk write iops: 3466.5
disk write throughput: 460.1 MiB/s
insert iops: 6927.5/s
insert throughput: 433.0 MiB/s
insert lat p50: 37us
insert lat p90: 85us
insert lat p99: 425us
insert lat p999: 1079us
insert lat p9999: 1839us
insert lat p99999: 5631us
insert lat pmax: 11839us
get iops: 16360.6/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1071us
get hit lat p90: 2191us
get hit lat p99: 3359us
get hit lat p999: 6111us
get hit lat p9999: 13567us
get hit lat p99999: 23679us
get hit lat pmax: 24959us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

### PR (run 1)

```
Total:
disk total iops: 3846.9
disk total throughput: 486.0 MiB/s
disk read iops: 372.1
disk read throughput: 24.7 MiB/s
disk write iops: 3474.8
disk write throughput: 461.3 MiB/s
insert iops: 6944.5/s
insert throughput: 434.0 MiB/s
insert lat p50: 38us
insert lat p90: 102us
insert lat p99: 361us
insert lat p999: 1031us
insert lat p9999: 1783us
insert lat p99999: 2207us
insert lat pmax: 3055us
get iops: 16361.7/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1159us
get hit lat p90: 2399us
get hit lat p99: 3503us
get hit lat p999: 7199us
get hit lat p9999: 12159us
get hit lat p99999: 21119us
get hit lat pmax: 22399us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

### PR (run 2)

```
Total:
disk total iops: 3841.4
disk total throughput: 484.6 MiB/s
disk read iops: 381.1
disk read throughput: 25.3 MiB/s
disk write iops: 3460.2
disk write throughput: 459.3 MiB/s
insert iops: 6915.4/s
insert throughput: 432.2 MiB/s
insert lat p50: 38us
insert lat p90: 104us
insert lat p99: 499us
insert lat p999: 1247us
insert lat p9999: 2191us
insert lat p99999: 2703us
insert lat pmax: 4351us
get iops: 16365.4/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1535us
get hit lat p90: 3087us
get hit lat p99: 4047us
get hit lat p999: 6463us
get hit lat p9999: 12671us
get hit lat p99999: 20863us
get hit lat pmax: 21887us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```