@MrCroxx 2024-04-18
## flamegraph

### cmd

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn flamegraph --no-inline -- ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 204800 --region-size 64 --lookup-range 10000 --flushers 12 --reclaimers 12 --time 60 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --io-size 262144 --distribution zipf --distribution-zipf-s 0.8
```

### serde

![foyer serde](https://github.com/MrCroxx/foyer/assets/22407295/a96728bc-9b38-40ba-8452-3e530190bf73)

### main

![foyer main](https://github.com/MrCroxx/foyer/assets/22407295/d3a7fc8e-b473-4d65-900a-86211b3a5228)