@MrCroxx 2024-04-18
serde:

<img width="1920" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/5636ab8c-e6e1-4a3c-8d96-b77f259a1af7">

main:

<img width="1920" alt="image" src="https://github.com/MrCroxx/foyer/assets/22407295/86ac3ab9-bd5c-4f6c-b464-388ed91b633c">

Seems serde bincode cannot fully utilize simd. Let me check if there are some related features to enable.