@MrCroxx 2024-04-18
Looks much better with `serde_bytes`, as @wangrunji0408 recommended.

![foyer serde bytes](https://github.com/MrCroxx/foyer/assets/22407295/b874a859-dfed-43e3-bdb3-fc144cd1f403)

```
Total:
disk total iops: 3751.2
disk total throughput: 478.7 MiB/s
disk read iops: 288.1
disk read throughput: 19.1 MiB/s
disk write iops: 3463.0
disk write throughput: 459.6 MiB/s
insert iops: 6919.8/s
insert throughput: 432.5 MiB/s
insert lat p50: 35us
insert lat p90: 72us
insert lat p99: 515us
insert lat p999: 1159us
insert lat p9999: 2303us
insert lat p99999: 6783us
insert lat pmax: 8383us
get iops: 16359.1/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1087us
get hit lat p90: 2351us
get hit lat p99: 3375us
get hit lat p999: 5727us
get hit lat p9999: 13439us
get hit lat p99999: 23679us
get hit lat pmax: 24191us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```